refactor: extract color conversion logic to dedicated c64u-color module
Separates VIC-II color palette and conversion logic into focused module:

New Files:
• src/c64u-color.h - Color conversion API and documentation
• src/c64u-color.c - VIC-II palette and optimized conversion implementation

Refactored:
• Moved vic_colors[16] palette definition to c64u-color.c
• Extracted init_color_conversion_lut() and convert_pixels_optimized()
• Updated includes in c64u-video.c and c64u-source.c
• Cleaned up c64u-video.h interface to focus on video operations
• Added c64u-color.c to CMakeLists.txt build configuration

Benefits:
• Improved modularity and separation of concerns
• Color logic isolated for easier testing and optimization
• Cleaner video module focused on frame assembly and processing
• Reusable color conversion for potential future modules

All existing functionality preserved with zero performance impact.
