+++ CMakeLists.txt
@@ -42,6 +42,7 @@ target_sources(
     src/c64u-network.c
     src/c64u-protocol.c
     src/c64u-video.c
+    src/c64u-color.c
     src/c64u-audio.c
     src/c64u-source.c
     src/c64u-record.c
