+++ src/c64u-source.c
@@ -11,6 +11,7 @@
 #include "c64u-types.h"
 #include "c64u-protocol.h"
 #include "c64u-video.h"
+#include "c64u-color.h"
 #include "c64u-network.h"
 #include "c64u-audio.h"
 #include "c64u-record.h"
